% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{A dataset containing the fatal injuries within the USA for 2013}
\format{A data frame with x rows and y variables}
\usage{
data(accident_2013)
}
\description{
\itemize{
  \item 1
  \item 2
  \item 3
  \item etcetera
}
}
\keyword{datasets}
