% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string denoting the (compressed) file name}
}
\value{
This function returns the input data in a data frame table
}
\description{
This function reads a (compressed) csv file with data of fatal injuries in motor vehicle traffic crashes in the USA.
}
\note{
This function will give a message "file <filename> does not exist" when the input file does not exist
}
\examples{
\dontrun{
fars_read("data/accident_2013.csv")
fars_read("data/accident_2013.csv.bz2")
fars_read("data/accident_2013.csv.zip")
}

}
