% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An array of or single integer or character string denoting an array of years or a single year}
}
\value{
This function returns a list with repeating the month and per month the year
}
\description{
This function uses the data file with fatal injuries in motor vehicle traffic crashes in the USA based on the input and selects
all the months (via the MONTH field) contained in the data file with the year(s), which is/are used as input.
}
\note{
This function will give a message "invalid year: <year>" when the input file based on the input cannot be processed
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years("2014")
fars_read_years(c(2013,2014))
}

}
