% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An array of or single integer or character string denoting an array of years or a single year}
}
\value{
This function returns a list with per month the total casualties per (input) year
}
\description{
This function uses the data file with fatal injuries in motor vehicle traffic crashes in the USA to summarizes the total number 
of fatal injuries per month/year.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years("2014")
fars_summarize_years(c(2013,2014))
}

}
