% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or character string denoting a year}
}
\value{
This function returns the character string "accident_<input as numeric>.csv.bz2" 
        with <input as numeric> the given input converted to numeric
}
\description{
This function creates, with the given input, the name of the data file with fatal injuries in motor vehicle traffic crashes in the USA.
}
\examples{
make_filename(2013)
make_filename("2014")

}
